# zstd: translate fseDecoder.buildDtable into asm